Allow for weight to be a decimal

+++ ItemGen/Ammunition.cs
@@ -67,7 +67,7 @@ namespace ItemGen
             objXmlItem.ReadString("Availability", ref Avail);
             objXmlItem.ReadInt("Cost", ref Cost);
             objXmlItem.ReadInt("DeviceRating", ref DeviceRating);
-            objXmlItem.ReadInt("Weight", ref Weight);
+            objXmlItem.ReadDouble("Weight", ref Weight);
 
             //Load Armor Info
             this.itemType = ItemType.Ammunition;
